--- src/pages/loginPage/components/styles/LoginForm.module.scss
@@ -1,78 +1,0 @@
-@use '../../../../styles/variables' as *;
-@use '../../../../styles/mixins' as m;
-
-.LoginFormContainer {
-  @include m.flex-column-center(center, center, null);
-  > *:first-child {
-    //  첫 번째 자식만
-    margin-bottom: 20px;
-  }
-  .inputField {
-    @include m.flex-column-center(flex, flex, 10px);
-    width: 100%;
-    max-width: min(83vw, 365.5px);
-    .inputLabel {
-      @include m.font-size-weight(14px, 500, 1.4);
-      color: $main-color2;
-      padding-left: 8px;
-    }
-  }
-
-  .emailWrapper {
-    @include m.flex-row-center(center, center, 6px);
-    padding: 18px 20px;
-    height: 60px;
-    border: 1px solid $gray-200;
-    border-radius: 12px;
-  }
-
-  .passwordWrapper {
-    @include m.flex-row-center(center, center, 6px);
-    padding: 18px 20px;
-    height: 60px;
-    border: 1px solid $gray-200;
-    border-radius: 12px;
-    .togglePasswordBtn {
-      width: 16px;
-      border: none;
-      background: none;
-      outline: none;
-      display: flex;
-      align-items: center;
-    }
-  }
-
-  .inputBox {
-    border: none;
-    background: none;
-    outline: none;
-    width: 100%;
-    height: 100%;
-    @include m.font-size-weight(14px, 400, 22px);
-    color: $main-color2;
-  }
-
-  &::placeholder {
-    @include m.font-size-weight(14px, 400, 22px);
-    color: #777c89;
-  }
-
-  .idSaveContainer {
-    margin-top: 12px;
-    width: 100%;
-    max-width: min(81.4vw, 355px);
-    .idSaveBtn {
-      @include m.flex-row-center(flex, center, 6px);
-      @include m.font-size-weight(14px, 400, 22px);
-      color: #9ea3b2;
-      border: none;
-      background: none;
-      outline: none;
-    }
-  }
-  .loginBtnContainer {
-    margin-top: 66px;
-    width: 100%;
-    max-width: min(83vw, 365.5px);
-  }
-}
